moved post concern into idea model
