+++ back/app/models/idea.rb
@@ -59,14 +59,17 @@
 #  fk_rails_...  (project_id => projects.id)
 #
 class Idea < ApplicationRecord
-  include Post
+  include PgSearch::Model
+  include GeoJsonHelpers
   include AnonymousParticipation
   extend OrderAsSpecified
 
-  before_save :convert_wkt_geo_custom_field_values_to_geojson
+  PUBLICATION_STATUSES = %w[draft submitted published].freeze
+  SUBMISSION_STATUSES = %w[submitted published].freeze
 
   slug from: proc { |idea| idea.participation_method_on_creation.generate_slug(idea) }
 
+  belongs_to :author, class_name: 'User', optional: true
   belongs_to :project, touch: true
   belongs_to :creation_phase, class_name: 'Phase', optional: true
   belongs_to :idea_status, optional: true
@@ -89,6 +92,21 @@ class Idea < ApplicationRecord
 
   belongs_to :assignee, class_name: 'User', optional: true
   belongs_to :manual_votes_last_updated_by, class_name: 'User', optional: true
+
+  has_many :activities, as: :item
+
+  has_many :comments, dependent: :destroy
+  has_many :internal_comments, dependent: :destroy
+
+  has_many :reactions, as: :reactable, dependent: :destroy
+  has_many :likes, -> { where(mode: 'up') }, as: :reactable, class_name: 'Reaction'
+  has_many :dislikes, -> { where(mode: 'down') }, as: :reactable, class_name: 'Reaction'
+  has_one :user_reaction, ->(user_id) { where(user_id: user_id) }, as: :reactable, class_name: 'Reaction'
+
+  has_many :spam_reports, as: :spam_reportable, class_name: 'SpamReport', dependent: :destroy
+
+  before_destroy :remove_notifications # Must occur before has_many :notifications (see https://github.com/rails/rails/issues/5205)
+  has_many :notifications, dependent: :nullify
 
   has_many :ideas_topics, dependent: :destroy
   has_many :topics, -> { order(:ordering) }, through: :ideas_topics
@@ -117,6 +135,8 @@ class Idea < ApplicationRecord
     post.after_validation :set_assigned_at, if: ->(record) { record.assignee_id && record.assignee_id_changed? }
   end
 
+  validates :publication_status, presence: true, inclusion: { in: PUBLICATION_STATUSES }
+
   with_options if: :supports_built_in_fields? do
     validates :title_multiloc, presence: true, multiloc: { presence: true }
     validates :body_multiloc, presence: true, multiloc: { presence: true, html: true }
@@ -138,11 +158,27 @@ class Idea < ApplicationRecord
     before_validation :sanitize_body_multiloc, if: :body_multiloc
   end
 
+  before_save :convert_wkt_geo_custom_field_values_to_geojson
+
   after_update :fix_comments_count_on_projects
 
   pg_search_scope :search_by_all,
     against: %i[title_multiloc body_multiloc custom_field_values slug],
     using: { tsearch: { prefix: true } }
+
+  pg_search_scope :restricted_search,
+    against: %i[title_multiloc body_multiloc slug],
+    using: { tsearch: { prefix: true } }
+
+  pg_search_scope :search_any_word,
+    against: %i[title_multiloc body_multiloc],
+    using: { tsearch: { any_word: true } }
+
+  # Note from: https://github.com/Casecommons/pg_search
+  # > Searching through associations
+  # > It is possible to search columns on associated models. Note that if you do this,
+  # > it will be impossible to speed up searches with database indexes. However, it
+  # > is supported as a quick way to try out cross-model searching.
 
   scope :with_some_topics, (proc do |topics|
     ideas = joins(:ideas_topics).where(ideas_topics: { topic: topics })
@@ -158,6 +194,15 @@ class Idea < ApplicationRecord
     joins(project: [:admin_publication])
       .where(projects: { admin_publications: { publication_status: publication_status } })
   end)
+
+  scope :order_new, ->(direction = :desc) { order(published_at: direction) }
+  scope :order_random, lambda { |user|
+    hash_part_for_today_and_user = Time.zone.today.to_s + user&.id.to_s
+    order(Arel.sql("md5(concat(#{table_name}.id, '#{hash_part_for_today_and_user}'))"))
+  }
+  scope :order_author_name, lambda { |direction = :desc|
+    includes(:author).order('users.first_name' => direction, 'users.last_name' => direction)
+  }
 
   scope :order_popular, ->(direction = :desc) { order(Arel.sql("(likes_count - dislikes_count) #{direction}, ideas.id")) }
   # based on https://medium.com/hacking-and-gonzo/how-hacker-news-ranking-algorithm-works-1d9b0cf2c08d
@@ -215,6 +260,15 @@ class Idea < ApplicationRecord
     where(id: idea_ids)
   }
 
+  scope :with_bounding_box, (proc do |coordinates|
+    x1, y1, x2, y2 = JSON.parse(coordinates)
+    where('ST_Intersects(ST_MakeEnvelope(?, ?, ?, ?), location_point)', x1, y1, x2, y2)
+  end)
+
+  scope :draft, -> { where(publication_status: 'draft') }
+  scope :published, -> { where publication_status: 'published' }
+  scope :submitted_or_published, -> { where publication_status: SUBMISSION_STATUSES }
+
   def just_submitted?
     # It would be better to foresee separate endpoints for submission,
     # rather than relying on Rails dirty to detect publication.
@@ -270,6 +324,26 @@ class Idea < ApplicationRecord
     self.manual_votes_amount = amount
     self.manual_votes_last_updated_by = user if user
     self.manual_votes_last_updated_at = Time.now
+  end
+
+  def draft?
+    publication_status == 'draft'
+  end
+
+  def submitted_or_published?
+    SUBMISSION_STATUSES.include? publication_status
+  end
+
+  def published?
+    publication_status == 'published'
+  end
+
+  def score
+    likes_count - dislikes_count
+  end
+
+  def author_name
+    @author_name ||= author&.full_name
   end
 
   private
@@ -382,6 +456,32 @@ class Idea < ApplicationRecord
       phase.pmethod.supports_input_term?
     end
     future_phases_with_input_term.first&.input_term
+  end
+
+  def strip_title
+    title_multiloc.each do |key, value|
+      title_multiloc[key] = value.strip
+    end
+  end
+
+  def set_submitted_at
+    self.submitted_at ||= Time.zone.now
+  end
+
+  def set_published_at
+    self.published_at ||= Time.zone.now
+  end
+
+  def set_assigned_at
+    self.assigned_at ||= Time.zone.now
+  end
+
+  def remove_notifications
+    notifications.each do |notification|
+      unless notification.update idea: nil
+        notification.destroy!
+      end
+    end
   end
 end
